requirements.txt
requests
pandas
psycopg2-binary
sqlalchemy